+++ CHANGELOG.md
@@ -323,3 +323,5 @@ Version 0.2.0 introduces the core media library functionality with automatic USB
 - [2025-11-02T18:27:56.4449542Z] Completed tasks: M2.7
 
 - [2025-11-02T18:32:02.8828098Z] Completed tasks: M2.8
+
+- [2025-11-02T18:34:56.0415483Z] Completed tasks: M2.9